requirements.txt
msal>=1.24.0
databricks-sdk>=0.12.0

